Created the frame for adding posts to a database
I used the sqlite3 python libraries and simply create a database, poll
using the RedditGrabber class, then add them to that database. At the
moment it will just delete the .db on every run.

Also added a "rank" to the posts so their order is more concrete when it
is saved.