Classed out RedditRecord and decoupled from grab
RedditRecord and RedditGrab are now independent from each other.

RedditRecord now takes a list of tuples containing the data to be logged
so a parent function can use the two separately.

I took the time to ensure that the wrong input into the database
would raise errors so no wonky values can be committed.
There is a pseudotest for the wrong values as the main function of
RedditRecord.py as well as a set of good values.