Added graphing to display frontpage frequency
displaygraph.py will show a graph which displays the amount of times a
subreddit has made it to the top 20 on the front page based on the
database created (and polled every hour) in RedditMonitor. That database
must exist before displaygraph.py will be relevant.

RedditRecorder was modified to output dictionaries based on both the
poll_time -> [post, post...] and the subreddit -> [headline, headline..]